Recursive Option + Hidden Hidden Files
Added a check to hide hidden files from the list output, which will make it a bit easier to find the right kinds of files I'm looking for, specifically the library's song files.

Added options to disable the previously listed addition, and the existing recursive behavior of getting the files within child directories. I also added some JSDoc documentation to describe the function a bit, too.

https://stackoverflow.com/questions/50121881/node-js-recursively-list-full-path-of-files